requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepdiff==8.0.1
Deprecated==1.2.14
distro==1.9.0
emoji==2.12.1
exceptiongroup==1.2.2
fastapi==0.114.2
filelock==3.16.0
filetype==1.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
googleapis-common-protos==1.65.0
googlesearch-python==1.2.5
greenlet==3.1.0
groq==0.11.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.40
langchain-groq==0.1.9
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.121
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orderly-set==5.2.2
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.0.2
pillow==10.4.0
posthog==3.6.6
protobuf==4.25.4
psycopg2-binary==2.9.10
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.5.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-iso639==2024.4.27
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
selenium==4.21.0
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
starlette==0.38.5
streamlit==1.35.0
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.14.6
unstructured-client==0.25.8
urllib3==2.2.3
uvicorn==0.30.6
watchdog==5.0.2
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.1.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.11.1
zipp==3.20.2
